pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "good_music_bot"
version = "0.1.0"
description = "A discord music bot powered by discord.py"
readme = "README.md"

requires-python = ">=3.6"
dependencies = [
    "discord.py",
    "python-dotenv",
]